+++ src/kiara/pipeline/controller/batch.py
@@ -147,7 +147,7 @@ class BatchControllerManual(PipelineController):
 
                             traceback.print_stack()
                         log.error(
-                            f"Processing of step '{step_id}' from pipeline '{self.pipeline.structure.pipeline_id}' failed: {e}"
+                            f"Processing of step '{step_id}' from pipeline '{self.pipeline.title}' failed: {e}"
                         )
                         return False
                 self._processor.wait_for(*job_ids)
@@ -209,7 +209,7 @@ class BatchControllerManual(PipelineController):
 
                             traceback.print_stack()
                         log.error(
-                            f"Processing of step '{step_id}' from pipeline '{self.pipeline.structure.pipeline_id}' failed: {e}"
+                            f"Processing of step '{step_id}' from pipeline '{self.pipeline.title}' failed: {e}"
                         )
                         return False
                 self._processor.wait_for(*job_ids)